type cast
git-svn-id: svn+ssh://frodo/cluster/svn/projects@20548 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ pfl/ctlsvr.c
@@ -1360,7 +1360,7 @@ psc_ctlparam_opstats(int fd, struct psc_ctlmsghdr *mh,
 
 	reset = (mh->mh_type == PCMT_SETPARAM);
 
-	for (i = 0; i < nitems(pflctl_opstats); i++) {
+	for (i = 0; i < (int)nitems(pflctl_opstats); i++) {
 		pos = &pflctl_opstats[i];
 		if (pos->pos_name == NULL)
 			continue;
